Validations done

+++ StudentDashBoard/src/main/java/com/hostbooks/service/impl/StudentServiceImpl.java
@@ -2,11 +2,16 @@ package com.hostbooks.service.impl;
 
 import java.util.List;
 
+import javax.validation.ConstraintViolationException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.hostbooks.dao.StudentDao;
+import com.hostbooks.exceptions.DataNotFoundException;
+import com.hostbooks.exceptions.IdNotFoundException;
+import com.hostbooks.exceptions.MailNotValidException;
 import com.hostbooks.model.Student;
 import com.hostbooks.service.StudentService;
 
@@ -18,29 +23,54 @@ public class StudentServiceImpl implements StudentService {
 
 	@Transactional
 	public Student saveStudent(Student student) {
+		
+		String mail=student.getStudentMail();
+		if(mail==null) {
+			throw new MailNotValidException("Please Enter Valid Email");
+		}
 		return studentDao.saveStudent(student);
 	}
 
-	@Transactional
+	@Transactional(readOnly = true)
 	public List<Student> getStudents() {
-		return studentDao.getAllStudents();
-
+		List<Student> allStudents = studentDao.getAllStudents();
+		if (allStudents == null) {
+			throw new DataNotFoundException();
+		}
+		return allStudents;
 	}
 
 	@Transactional
 	public void deleteStudent(Integer id) {
+
+		Student studentByID = studentDao.getStudentByID(id);
+		if (studentByID == null) {
+			throw new IdNotFoundException("Id  not found");
+		}
+
+		/*
+		 * if (id == null) { throw new IdNotFoundException(""); }
+		 */
 		studentDao.deleteStudent(id);
 	}
 
-	@Transactional
+	@Transactional(readOnly = true)
 	public Student getStudentByID(Integer id) {
+
+		Student studentByID = studentDao.getStudentByID(id);
+		if (studentByID == null) {
+			throw new IdNotFoundException("Id  not found");
+		}
 		return studentDao.getStudentByID(id);
 
 	}
 
 	@Transactional
 	public Student updateStudent(Student student) {
-
+		Integer id = student.getStudentId();
+		if (id == null) {
+			throw new IdNotFoundException("");
+		}
 		return studentDao.updateStudent(student);
 	}
